Changed csite%soil_water to an 8-byte floating point vector    Most modifications revolving around this were related to passing    the correct argument type, hence changes to decomposition,phenology_driv,    photosyn_driv.  IO statements were changed to include a double precision    variable,however, it was noticed that HDF files are storing reals with 32    bit precision anyway, this is something that should be looked into, becuase    fortran native precision for reals is only 16 bits.  This would halve the    size of the files if it is true.  Anyway, I added a new type to the list    of real, character and integer IO.    One subroutine in therm_lib, qwtk(), was mirrored. qwtk() does some thermo    with water input, but it does not always do a calculation on soil_water    as the argument, sometimes it is surface or leaf water. So the mirror    subroutine is called qwtk8(). In all instances where soil water is an    argument, now qwtk8() is called.

+++ ED/src/dynamics/phenology_driv.f90
@@ -261,7 +261,7 @@ subroutine phenology_thresholds(daylight, soil_temp, soil_water, soil_class,  &
 
   real, intent(in) :: daylight
   real, intent(in) :: soil_temp
-  real, dimension(nzg), intent(in) :: soil_water
+  real(kind=8), dimension(nzg), intent(in) :: soil_water
   integer, dimension(nzg), intent(in) :: soil_class
   real, intent(inout) :: sum_dgd
   real, intent(inout) :: sum_chd
